refactor: move WorkspaceStateKey/GlobalStateKey to @shared/state/stateKeys
The state key enums were defined in @common/state/stateKeys, but @common is
documented as "backend-only helpers" (imports vscode). Because the keys are
pure string constants, vscode-free zones (src/agent/, src/latex/, src/model/,
src/tools/) had to import from the sub-file path rather than the @common/state
barrel to avoid accidentally pulling in the vscode module at load time — a
brittle, comment-enforced workaround.

The canonical home for cross-cutting vscode-free constants is @shared, which
explicitly prohibits vscode imports and is consumed by all hosts (extension,
desktop, CLI). Moving the enums there makes the correct import path self-evident:
any zone that needs state keys imports from @shared/state/stateKeys.

Changes:
- Add src/shared/state/stateKeys.ts with the WorkspaceStateKey/GlobalStateKey
  enums and INSTRUCTION_PREFIX constant
- Replace src/common/state/stateKeys.ts with a re-export shim so the
  @common/state barrel and any remaining @common/state/stateKeys direct
  imports remain working without a touch
- Update @shared/state/index.ts to re-export from the new file
- Update 21 import sites across src/ and packages/ to use @shared/state/stateKeys
