+++ src/agent/features/registerMemory.ts
@@ -1,6 +1,6 @@
 import { getGlobalState } from '@agent/core/stateStore';
 import { registerToolInjection } from '@agent/runtime/toolInjection';
-import { GlobalStateKey } from '@common/state/stateKeys';
+import { GlobalStateKey } from '@shared/state/stateKeys';
 
 export function registerMemoryFeature(): void {
   registerToolInjection({
