+++ src/agent/output/compileCheck.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import type { AgentTrace } from '@agent/trace';
 import { getWorkspaceState } from '@agent/core/stateStore';
 import { toErrorMessage } from '@common/errors';
-import { WorkspaceStateKey } from '@common/state/stateKeys';
+import { WorkspaceStateKey } from '@shared/state/stateKeys';
 import { hasLatexCompiler } from '@latex/latexToolchain';
 import { compileLatex2Pdf } from '@latex/texTools';
 import type {
