+++ src/agent/runtime/ModelFactory.ts
@@ -7,7 +7,7 @@ import { ModelHandler } from '@agent/modelHandlers/ModelHandler';
 import type { ProviderMessage } from '@agent/modelHandlers/types/ProviderMessage';
 import { getGlobalState } from '@agent/core/stateStore';
 import { getConfig } from '@utils/config/configUtils';
-import { GlobalStateKey } from '@common/state/stateKeys';
+import { GlobalStateKey } from '@shared/state/stateKeys';
 import * as logger from '@logger/logUtils';
 import { getUseOpenRouter } from '@utils/config/providerConfig';
 import { LEVEL_TO_EFFORT } from './reasoningEffort';
