+++ src/agent/runtime/delegationPolicy.ts
@@ -9,7 +9,7 @@
 import { getExecutionStore } from '@agent/storage';
 import type { ExecutionMeta } from '@agent/storage/ExecutionKVStore';
 import { getWorkspaceState } from '@agent/core/stateStore';
-import { WorkspaceStateKey } from '@common/state/stateKeys';
+import { WorkspaceStateKey } from '@shared/state/stateKeys';
 import type { ExecutionId } from '@shared/schemas';
 import {
   clampNestedDelegationDepth,
