+++ src/agent/runtime/helperModel.ts
@@ -11,7 +11,7 @@ import { MODEL_CONFIGS } from 'llm-zoo';
 import type { ModelHandler } from '@agent/modelHandlers/ModelHandler';
 import { createModelHandler } from '@agent/runtime/ModelFactory';
 import { getGlobalState } from '@agent/core/stateStore';
-import { GlobalStateKey } from '@common/state/stateKeys';
+import { GlobalStateKey } from '@shared/state/stateKeys';
 import { getModelUnavailableReason } from '@model/computeModelOptions';
 import { DEFAULT_HELPER_MODEL } from '@shared/constants/providers';
 import { isNonEmptyString } from '@utils/core';
