+++ src/latex/latexdiff.ts
@@ -5,7 +5,7 @@ import { z } from 'zod';
 import { extractLastRoundMatch } from '@agent/utils/mergeFileUtils';
 import { tryGetWorkspaceState } from '@agent/core/stateStore';
 import { formatError, toErrorMessage } from '@common/errors';
-import { WorkspaceStateKey } from '@common/state/stateKeys';
+import { WorkspaceStateKey } from '@shared/state/stateKeys';
 import * as logger from '@logger/logUtils';
 import { LATEX_CONFIG_DEFAULTS } from '@shared/constants/latex';
 import { flexibleFS, pathToLocation, type FileLocation } from '@utils/files';
