+++ src/shared/state/index.ts
@@ -5,3 +5,8 @@ export {
   createWebviewStorage,
   type StateStorage,
 } from './PersistedState';
+export {
+  WorkspaceStateKey,
+  GlobalStateKey,
+  INSTRUCTION_PREFIX,
+} from './stateKeys';
